+++ tests/testthat/test-as_kable_extra.R
@@ -60,3 +60,18 @@ test_that("tbl_merge/tbl_stack", {
   expect_error(tbl_stack_ex1 %>% as_kable_extra(), NA)
   expect_warning(tbl_stack_ex1 %>% as_kable_extra(), NA)
 })
+
+test_that("indent2", {
+  expect_error(
+    trial %>%
+      select(age) %>%
+      tbl_summary() %>%
+      modify_table_styling(
+        columns = label,
+        rows = variable == "age" & row_type != "label",
+        text_format = "indent2"
+      ) %>%
+      as_kable_extra(),
+    NA
+  )
+})